EANX_CC/UPCA_CC: fix crash in `dbar_date()` on not checking length   in `cc_binary_string()`, ticket #300 (#5 & #6), props Andre Maute;   add other checks for length on processing encoding mode PDF417: fix out-of-bounds crash on overrunning string and codeword   buffers by tripling size (convert to `short` instead of `int` to   guard against too much stack), ticket #300 (#7 & #10), props Andre   Maute; (TODO: add some checks instead to bail out earlier?) CODEONE: fix looping on latch crash in `c1_encode()`, ticket #300 (#8),   props Andre Maute CODABLOCKF: fix crash on negative overflow of `columns` (`option_2`),   ticket #300 (#9), props Andre Maute library: add `debug_print_escape()` helper for ZINT_DEBUG_PRINT

+++ backend/pdf417.h
@@ -56,7 +56,7 @@ INTERNAL_DATA_EXTERN const unsigned short pdf_rap_side[52];
 /* Centre Row Address Pattern from Table 2 */
 INTERNAL_DATA_EXTERN const unsigned short pdf_rap_centre[52];
 
-INTERNAL void pdf_byteprocess(int *chainemc, int *p_mclength, const unsigned char chaine[], int start,
+INTERNAL void pdf_byteprocess(short *chainemc, int *p_mclength, const unsigned char chaine[], int start,
                 const int length, const int lastmode, const int debug);
 
 /* vim: set ts=4 sw=4 et : */
